Rename preview.py to core.py
